convert Home page to TypeScript
